Rename SectionHeaderComponent to LabelComponent
