+++ Apps/WebClient/src/ClientApp/src/components/common/LabelComponent.vue
